+++ docs/conf.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# scrapy-crawlera documentation build configuration file, created by
+# scrapy-zyte-smartproxy documentation build configuration file, created by
 # sphinx-quickstart on Sat Jan 21 13:17:41 2017.
 #
 # This file is execfile()d with the current directory set to its
@@ -54,9 +54,9 @@ source_suffix = '.rst'
 master_doc = 'index'
 
 # General information about the project.
-project = u'scrapy-crawlera'
-copyright = u'2011-2017, Scrapinghub'
-author = u'Scrapinghub'
+project = u'scrapy-zyte-smartproxy'
+copyright = u'2011-2021, Zyte Group Ltd'
+author = u'Zyte'
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the
@@ -64,9 +64,9 @@ author = u'Scrapinghub'
 #
 
 try:
-    import scrapy_crawlera
-    version = '.'.join(scrapy_crawlera.__version__.split('.')[:2])
-    release = scrapy_crawlera.__version__
+    import scrapy_zyte_smartproxy
+    version = '.'.join(scrapy_zyte_smartproxy.__version__.split('.')[:2])
+    release = scrapy_zyte_smartproxy.__version__
 except ImportError:
     version = ''
     release = ''
@@ -111,7 +111,7 @@ todo_include_todos = False
 # -- Options for HTMLHelp output ------------------------------------------
 
 # Output file base name for HTML help builder.
-htmlhelp_basename = 'scrapy-crawleradoc'
+htmlhelp_basename = 'scrapy-zyte-smartproxydoc'
 
 
 # -- Options for LaTeX output ---------------------------------------------
@@ -138,8 +138,13 @@ latex_elements = {
 # (source start file, target name, title,
 #  author, documentclass [howto, manual, or own class]).
 latex_documents = [
-    (master_doc, 'scrapy-crawlera.tex', u'scrapy-crawlera Documentation',
-     u'Scrapinghub', 'manual'),
+    (
+        master_doc,
+        'scrapy-zyte-smartproxy.tex',
+        u'scrapy-zyte-smartproxy Documentation',
+        u'Zyte',
+        'manual',
+    ),
 ]
 
 
@@ -148,7 +153,7 @@ latex_documents = [
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
 man_pages = [
-    (master_doc, 'scrapy-crawlera', u'scrapy-crawlera Documentation',
+    (master_doc, 'scrapy-zyte-smartproxy', u'scrapy-zyte-smartproxy Documentation',
      [author], 1)
 ]
 
@@ -159,8 +164,8 @@ man_pages = [
 # (source start file, target name, title, author,
 #  dir menu entry, description, category)
 texinfo_documents = [
-    (master_doc, 'scrapy-crawlera', u'scrapy-crawlera Documentation',
-     author, 'scrapy-crawlera', 'One line description of project.',
+    (master_doc, 'scrapy-zyte-smartproxy', u'scrapy-zyte-smartproxy Documentation',
+     author, 'scrapy-zyte-smartproxy', 'One line description of project.',
      'Miscellaneous'),
 ]