+++ scrapy_zyte_smartproxy/middleware.py
@@ -10,16 +10,16 @@ from scrapy.resolver import dnscache
 from scrapy.exceptions import ScrapyDeprecationWarning
 from twisted.internet.error import ConnectionRefusedError, ConnectionDone
 
-from scrapy_crawlera.utils import exp_backoff
+from scrapy_zyte_smartproxy.utils import exp_backoff
 
 
-class CrawleraMiddleware(object):
+class ZyteSmartProxyMiddleware(object):
 
-    url = 'http://proxy.crawlera.com:8010'
+    url = 'http://proxy.zyte.com:8011'
     maxbans = 400
     ban_code = 503
     download_timeout = 190
-    # Handle crawlera server failures
+    # Handle Zyte Smart Proxy Manager server failures
     connection_refused_delay = 90
     preserve_delay = False
     header_prefix = 'X-Crawlera-'
@@ -64,20 +64,25 @@ class CrawleraMiddleware(object):
             setattr(self, k, self._get_setting_value(spider, k, type_))
 
         self._fix_url_protocol()
-        self._headers = self.crawler.settings.get('CRAWLERA_DEFAULT_HEADERS', {}).items()
+        self._headers = self.crawler.settings.get('ZYTE_SMARTPROXY_DEFAULT_HEADERS', {}).items()
         self.exp_backoff = exp_backoff(self.backoff_step, self.backoff_max)
 
         if not self.enabled and not self.force_enable_on_http_codes:
             return
 
         if not self.apikey:
-            logging.warning("Crawlera can't be used without a APIKEY", extra={'spider': spider})
+            logging.warning(
+                "Zyte Smart Proxy Manager cannot be used without an API key",
+                extra={'spider': spider},
+            )
             return
 
         self._proxyauth = self.get_proxyauth(spider)
 
         logging.info(
-            "Using crawlera at %s (apikey: %s)" % (self.url, self.apikey[:7]),
+            "Using Zyte Smart Proxy Manager at %s (apikey: %s)" % (
+                self.url, self.apikey[:7]
+            ),
             extra={'spider': spider},
         )
 
@@ -85,8 +90,11 @@ class CrawleraMiddleware(object):
             # Setting spider download delay to 0 to get maximum crawl rate
             spider.download_delay = 0
             logging.info(
-                "CrawleraMiddleware: disabling download delays on Scrapy side to optimize delays introduced by Crawlera. "
-                "To avoid this behaviour you can use the CRAWLERA_PRESERVE_DELAY setting but keep in mind that this may slow down the crawl significantly",
+                "ZyteSmartProxyMiddleware: disabling download delays in "
+                "Scrapy to optimize delays introduced by Zyte Smart Proxy "
+                "Manager. To avoid this behaviour you can use the "
+                "ZYTE_SMARTPROXY_PRESERVE_DELAY setting, but keep in mind "
+                "that this may slow down the crawl significantly",
                 extra={'spider': spider},
             )
 
@@ -105,41 +113,41 @@ class CrawleraMiddleware(object):
     def _get_setting_value(self, spider, k, type_):
         if hasattr(spider, 'hubproxy_' + k):
             warnings.warn('hubproxy_%s attribute is deprecated, '
-                          'use crawlera_%s instead.' % (k, k),
+                          'use zyte_smartproxy_%s instead.' % (k, k),
                           category=ScrapyDeprecationWarning, stacklevel=1)
 
         if self.crawler.settings.get('HUBPROXY_%s' % k.upper()) is not None:
             warnings.warn('HUBPROXY_%s setting is deprecated, '
-                          'use CRAWLERA_%s instead.' % (k.upper(), k.upper()),
+                          'use ZYTE_SMARTPROXY_%s instead.' % (k.upper(), k.upper()),
                           category=ScrapyDeprecationWarning, stacklevel=1)
 
         o = getattr(self, k, None)
         s = self._settings_get(
-            type_, 'CRAWLERA_' + k.upper(), self._settings_get(
+            type_, 'ZYTE_SMARTPROXY_' + k.upper(), self._settings_get(
                 type_, 'HUBPROXY_' + k.upper(), o))
         return getattr(
-            spider, 'crawlera_' + k, getattr(spider, 'hubproxy_' + k, s))
+            spider, 'zyte_smartproxy_' + k, getattr(spider, 'hubproxy_' + k, s))
 
     def _fix_url_protocol(self):
         if self.url.startswith('https://'):
-            logging.warning('CRAWLERA_URL "%s" set with "https://" protocol.' % self.url)
+            logging.warning('ZYTE_SMARTPROXY_URL "%s" set with "https://" protocol.' % self.url)
         elif not self.url.startswith('http://'):
-            logging.warning('Adding "http://" to CRAWLERA_URL %s' % self.url)
+            logging.warning('Adding "http://" to ZYTE_SMARTPROXY_URL %s' % self.url)
             self.url = 'http://' + self.url
 
     def is_enabled(self, spider):
         """Hook to enable middleware by custom rules."""
         if hasattr(spider, 'use_hubproxy'):
             warnings.warn('use_hubproxy attribute is deprecated, '
-                          'use crawlera_enabled instead.',
+                          'use zyte_smartproxy_enabled instead.',
                           category=ScrapyDeprecationWarning, stacklevel=1)
 
         if self.crawler.settings.get('HUBPROXY_ENABLED') is not None:
             warnings.warn('HUBPROXY_ENABLED setting is deprecated, '
-                          'use CRAWLERA_ENABLED instead.',
+                          'use ZYTE_SMARTPROXY_ENABLED instead.',
                           category=ScrapyDeprecationWarning, stacklevel=1)
         return (
-            getattr(spider, 'crawlera_enabled', self.crawler.settings.getbool('CRAWLERA_ENABLED')) or
+            getattr(spider, 'zyte_smartproxy_enabled', self.crawler.settings.getbool('ZYTE_SMARTPROXY_ENABLED')) or
             getattr(spider, 'use_hubproxy', self.crawler.settings.getbool("HUBPROXY_ENABLED"))
         )
 
@@ -148,19 +156,19 @@ class CrawleraMiddleware(object):
         return basic_auth_header(self.apikey, '')
 
     def process_request(self, request, spider):
-        from scrapy_crawlera import __version__
+        from scrapy_zyte_smartproxy import __version__
         if self._is_enabled_for_request(request):
-            self._set_crawlera_default_headers(request)
+            self._set_zyte_smartproxy_default_headers(request)
             request.meta['proxy'] = self.url
             request.meta['download_timeout'] = self.download_timeout
             request.headers['Proxy-Authorization'] = self._proxyauth
             if self.job_id:
                 request.headers['X-Crawlera-Jobid'] = self.job_id
-            request.headers['X-Crawlera-Client'] = 'scrapy-crawlera/%s' % __version__
-            self.crawler.stats.inc_value('crawlera/request')
-            self.crawler.stats.inc_value('crawlera/request/method/%s' % request.method)
+            request.headers['X-Crawlera-Client'] = 'scrapy-zyte-smartproxy/%s' % __version__
+            self.crawler.stats.inc_value('zyte_smartproxy/request')
+            self.crawler.stats.inc_value('zyte_smartproxy/request/method/%s' % request.method)
         else:
-            self._clean_crawlera_headers(request)
+            self._clean_zyte_smartproxy_headers(request)
 
     def _is_banned(self, response):
         return (
@@ -184,7 +192,7 @@ class CrawleraMiddleware(object):
         if not self._is_enabled_for_request(request):
             return self._handle_not_enabled_response(request, response)
 
-        if not self._is_crawlera_response(response):
+        if not self._is_zyte_smartproxy_response(response):
             return response
 
         key = self._get_slot_key(request)
@@ -197,17 +205,17 @@ class CrawleraMiddleware(object):
                 reason = 'autherror'
             self._set_custom_delay(request, next(self.exp_backoff), reason=reason)
         else:
-            self.crawler.stats.inc_value('crawlera/delay/reset_backoff')
+            self.crawler.stats.inc_value('zyte_smartproxy/delay/reset_backoff')
             self.exp_backoff = exp_backoff(self.backoff_step, self.backoff_max)
 
         if self._is_auth_error(response):
-            # When crawlera has issues it might not be able to authenticate users
-            # we must retry
-            retries = request.meta.get('crawlera_auth_retry_times', 0)
+            # When Zyte Smart Proxy Manager has issues it might not be able to
+            # authenticate users we must retry
+            retries = request.meta.get('zyte_smartproxy_auth_retry_times', 0)
             if retries < self.max_auth_retry_times:
                 return self._retry_auth(response, request, spider)
             else:
-                self.crawler.stats.inc_value('crawlera/retries/auth/max_reached')
+                self.crawler.stats.inc_value('zyte_smartproxy/retries/auth/max_reached')
                 logging.warning(
                     "Max retries for authentication issues reached, please check auth"
                     " information settings",
@@ -222,24 +230,24 @@ class CrawleraMiddleware(object):
                 after = response.headers.get('retry-after')
                 if after:
                     self._set_custom_delay(request, float(after), reason='banned')
-            self.crawler.stats.inc_value('crawlera/response/banned')
+            self.crawler.stats.inc_value('zyte_smartproxy/response/banned')
         else:
             self._bans[key] = 0
         # If placed behind `RedirectMiddleware`, it would not count 3xx responses
-        self.crawler.stats.inc_value('crawlera/response')
-        self.crawler.stats.inc_value('crawlera/response/status/%s' % response.status)
-        crawlera_error = response.headers.get('X-Crawlera-Error')
-        if crawlera_error:
-            self.crawler.stats.inc_value('crawlera/response/error')
+        self.crawler.stats.inc_value('zyte_smartproxy/response')
+        self.crawler.stats.inc_value('zyte_smartproxy/response/status/%s' % response.status)
+        zyte_smartproxy_error = response.headers.get('X-Crawlera-Error')
+        if zyte_smartproxy_error:
+            self.crawler.stats.inc_value('zyte_smartproxy/response/error')
             self.crawler.stats.inc_value(
-                'crawlera/response/error/%s' % crawlera_error.decode('utf8'))
+                'zyte_smartproxy/response/error/%s' % zyte_smartproxy_error.decode('utf8'))
         return response
 
     def process_exception(self, request, exception, spider):
         if not self._is_enabled_for_request(request):
             return
         if isinstance(exception, (ConnectionRefusedError, ConnectionDone)):
-            # Handle crawlera downtime
+            # Handle Zyte Smart Proxy Manager downtime
             self._clear_dns_cache()
             self._set_custom_delay(request, self.connection_refused_delay, reason='conn_refused')
 
@@ -250,20 +258,21 @@ class CrawleraMiddleware(object):
 
             retryreq = request.copy()
             retryreq.dont_filter = True
-            self.crawler.stats.inc_value('crawlera/retries/should_have_been_enabled')
+            self.crawler.stats.inc_value('zyte_smartproxy/retries/should_have_been_enabled')
             return retryreq
         return response
 
     def _retry_auth(self, response, request, spider):
         logging.warning(
-            "Retrying crawlera request for authentication issue",
+            "Retrying a Zyte Smart Proxy Manager request due to an "
+            "authentication issue",
             extra={'spider': self.spider},
         )
-        retries = request.meta.get('crawlera_auth_retry_times', 0) + 1
+        retries = request.meta.get('zyte_smartproxy_auth_retry_times', 0) + 1
         retryreq = request.copy()
-        retryreq.meta['crawlera_auth_retry_times'] = retries
+        retryreq.meta['zyte_smartproxy_auth_retry_times'] = retries
         retryreq.dont_filter = True
-        self.crawler.stats.inc_value('crawlera/retries/auth')
+        self.crawler.stats.inc_value('zyte_smartproxy/retries/auth')
         return retryreq
 
     def _clear_dns_cache(self):
@@ -284,7 +293,7 @@ class CrawleraMiddleware(object):
         parsed = urlparse(url)
         return parsed.netloc
 
-    def _is_crawlera_response(self, response):
+    def _is_zyte_smartproxy_response(self, response):
         return bool("X-Crawlera-Version" in response.headers)
 
     def _get_slot_key(self, request):
@@ -303,8 +312,8 @@ class CrawleraMiddleware(object):
             self._saved_delays[key] = slot.delay
         slot.delay = delay
         if reason is not None:
-            self.crawler.stats.inc_value('crawlera/delay/%s' % reason)
-            self.crawler.stats.inc_value('crawlera/delay/%s/total' % reason, delay)
+            self.crawler.stats.inc_value('zyte_smartproxy/delay/%s' % reason)
+            self.crawler.stats.inc_value('zyte_smartproxy/delay/%s/total' % reason, delay)
 
     def _restore_original_delay(self, request):
         """Restore original delay for slot if it was changed."""
@@ -314,23 +323,23 @@ class CrawleraMiddleware(object):
         if self._saved_delays[key] is not None:
             slot.delay, self._saved_delays[key] = self._saved_delays[key], None
 
-    def _clean_crawlera_headers(self, request):
+    def _clean_zyte_smartproxy_headers(self, request):
         """Remove X-Crawlera-* headers from the request."""
         targets = [
             header
             for header in request.headers
-            if self._is_crawlera_header(header)
+            if self._is_zyte_smartproxy_header(header)
         ]
         for header in targets:
             request.headers.pop(header, None)
 
-    def _is_crawlera_header(self, header_name):
+    def _is_zyte_smartproxy_header(self, header_name):
         if not header_name:
             return False
         header_name = header_name.decode('utf-8').lower()
         return header_name.startswith(self.header_prefix.lower())
 
-    def _set_crawlera_default_headers(self, request):
+    def _set_zyte_smartproxy_default_headers(self, request):
         for header, value in self._headers:
             if value is None:
                 continue
@@ -342,14 +351,17 @@ class CrawleraMiddleware(object):
             # Send a general warning once, and specific urls if LOG_LEVEL = DEBUG
             warnings.warn(
                 'The headers %s are conflicting on some of your requests. '
-                'Please check https://doc.scrapinghub.com/crawlera.html '
-                'for more information. You can set LOG_LEVEL=DEBUG to see the urls with problems'
+                'Please check '
+                'https://docs.zyte.com/smart-proxy-manager.html#request-headers '
+                'for more information. You can set LOG_LEVEL=DEBUG to see the '
+                'urls with problems.'
                 % str(self.conflicting_headers)
             )
             logging.debug(
                 'The headers %s are conflicting on request %s. X-Crawlera-UA '
-                'will be ignored. Please check https://doc.scrapinghub.com/cr'
-                'awlera.html for more information'
+                'will be ignored. Please check '
+                'https://docs.zyte.com/smart-proxy-manager.html#request-headers '
+                'for more information'
                 % (str(self.conflicting_headers), request.url),
                 extra={'spider': self.spider},
             )